move: V4AttnFfn and the attn-core runner/outputs into deepseek_v4_attn
Out of the @support_torch_compile model file; the model keeps one import.
Names lose the underscore now that they cross modules.
